+++ apps/chat/src/components/Chat/EmptyChatDescription.tsx
@@ -13,7 +13,7 @@ import {
   isReplayAsIsConversation,
 } from '@/src/utils/app/conversation';
 import { isEntityIdExternal } from '@/src/utils/app/id';
-import { getGroupMarketplaceEntityKey } from '@/src/utils/app/models';
+import { getGroupMarketplaceEntityKey } from '@/src/utils/app/marketplace';
 import { isEntityReadOnly } from '@/src/utils/app/permissions';
 
 import { Conversation } from '@/src/types/chat';
